Object code output identical.
Bit of a milestone: for the initial test source (the AVRVM up to the
interpret word) the pyavrasm assembler generates the same hex output at
the Amtel assembler.  Huzzah!

The only wrinkle is that due to the fact that the utility functions
high() and low() compute their return values during the first pass
instead of the second they were unable to correctly handle the forward
reference in the find word to the interpret word at the end of the
dictionary.

I fixed this temporarily by hand-coding the end-of-dictionary address
into the find word definition.  Good enough for now and the actual fix is
obvious.