Compiler that runs on a microcontroller.
I used the new Meta-II compiler from the previous commit and generated
(most of) a compiler that should run on the ATmega328P chip.

It needs some run-time support code, but I think I can reuse a lot of the
firmware code for that.

The interesting thing is that there's no point to compiling for the chip
*on* the chip.

Oh, eventually I do intend to have some on-board self-programming
support, but this compiler emits *assembly source* not binary code, so
even then it would need another "stage" or "pass" to create binary on
board.

It would be possible to rewrite the meta-compiler again to emit binary
directly but as this project is about teaching people how these things
work and not about how to create self-programming microcontrollers I
think I'll refrain.

It *will* be useful to use the AVR-targetting compiler to generate code
for things like the parser-style TWI (I2C) interface and other
VPRI-inspired goodness...