Config system with text bindings.
the config "system" is just execfile but it works nicely.

I set up a dict of text viewer bindings to apply and you can update that
in your config file to change how the TextViewer widget responds to
events.  It also makes it really easy to experiment with button bindings
for the functions that have before been tied just to the mouse chords.

It would be possible to let the config override the mouse chord bindings
using Tk's "meta-events" (like the '<<Modified>>' built-in event) but
it's more work than I want to do now, and I'm not sure how soon it would
be used.

And besides, the mouse chord integration with the interpreter is freakin'
awesome (IMHO.) ^_^