Arrange to commit after each change.
This is a little more involved and tightly-coupled than I would like but
it seems to work.

Still to do:

Notice when nothing has, in fact, changed and refrain from actually
commiting in that case.

Commit after the text widget manipulates the stack directly as opposed to
through the world.step() method.  The "right" way to do this would be to
make the text use commands, but it will be much easier for the moment to
just bunt and commit after (the existing) direct manipulation.