This is the original "AVR VM" rst file
...and its associated cruft.  I wanted a copy here so as to not "lose"
the additional functionality implemented in Amtel ASM before I get a
chance to recode it in Pigeon Computer ASM.

I'm reluctant to keep copies of the Amtel assembler and the chip defs
file so I'll delete them as soon as they're no longer needed.  (I'm sure
people will be able to get them from Amtel (or the internet) if needed
anyway. And on Linux the assembler needs Wine to run.  (One of lesser
reasons why I decided to implement my own assembler in Python.))